+++ src/iterator/raw.rs
@@ -9,12 +9,12 @@
 //! # Examples
 //!
 //! ```
-//! use embedded_graphics::{iterator::raw::RawDataSlice, pixelcolor::raw::{RawU16, BigEndian}};
+//! use embedded_graphics::{iterator::raw::RawDataSlice, pixelcolor::raw::{RawU16, BigEndianLsb0}};
 //!
 //! let data = [0xAA, 0xBB, 0x12, 0x34];
 //!
 //! // The data type and byte order needs to be specified explicitly to set the data format.
-//! let slice = RawDataSlice::<RawU16, BigEndian>::new(&data);
+//! let slice = RawDataSlice::<RawU16, BigEndianLsb0>::new(&data);
 //!
 //! let mut iter = slice.into_iter();
 //! assert_eq!(iter.next(), Some(RawU16::new(0xAABB)));
@@ -24,13 +24,11 @@
 //!
 //! [`ImageRaw`]: super::super::image::ImageRaw
 
-use core::{marker::PhantomData, slice};
+use core::marker::PhantomData;
 
-use byteorder::{ByteOrder, BE, LE};
+use embedded_graphics_core::pixelcolor::raw::DataOrder;
 
-use crate::pixelcolor::raw::{
-    BigEndian, LittleEndian, RawData, RawU1, RawU16, RawU2, RawU24, RawU32, RawU4, RawU8,
-};
+use crate::pixelcolor::raw::RawData;
 
 /// Raw data slice.
 ///
@@ -41,209 +39,85 @@ use crate::pixelcolor::raw::{
 /// [module-level documentation]: self
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "defmt", derive(::defmt::Format))]
-pub struct RawDataSlice<'a, R, BO> {
+pub struct RawDataSlice<'a, R, O> {
     data: &'a [u8],
     raw_type: PhantomData<R>,
-    byte_order: PhantomData<BO>,
+    data_order: PhantomData<O>,
 }
 
-impl<'a, R, BO> RawDataSlice<'a, R, BO> {
+impl<'a, R, O> RawDataSlice<'a, R, O> {
     /// Creates a new raw data slice.
     pub const fn new(data: &'a [u8]) -> Self {
         Self {
             data,
             raw_type: PhantomData,
-            byte_order: PhantomData,
+            data_order: PhantomData,
         }
     }
 }
 
-macro_rules! impl_bits_iterator {
-    ($type:ident, $bit_index_bits:expr) => {
-        impl<'a, BO> IntoIterator for RawDataSlice<'a, $type, BO> {
-            type Item = $type;
-            type IntoIter = BitsIterator<'a, $type>;
-
-            fn into_iter(self) -> Self::IntoIter {
-                BitsIterator::new(self.data)
-            }
-        }
-
-        impl<'a> Iterator for BitsIterator<'a, $type> {
-            type Item = $type;
-
-            #[inline]
-            fn next(&mut self) -> Option<Self::Item> {
-                self.data.get(self.index >> $bit_index_bits).map(|byte| {
-                    // Number of bits the value needs to be shifted to the right for the first pixel.
-                    let first_pixel_shift = 8 - $type::BITS_PER_PIXEL;
-
-                    // Index to one of the pixels inside this byte.
-                    let sub_index = self.index & (1 << $bit_index_bits) - 1;
-
-                    // Number of bits the value needs to be shifted.
-                    let shift = first_pixel_shift - (sub_index << 3 - $bit_index_bits);
-
-                    self.index += 1;
-
-                    $type::new(*byte >> shift)
-                })
-            }
-
-            #[inline]
-            fn nth(&mut self, n: usize) -> Option<Self::Item> {
-                self.index = self.index.saturating_add(n);
-                self.next()
-            }
-
-            fn size_hint(&self) -> (usize, Option<usize>) {
-                let size =
-                    (self.data.len() * (8 / $type::BITS_PER_PIXEL)).saturating_sub(self.index);
-
-                (size, Some(size))
-            }
-        }
-    };
-}
-
-impl_bits_iterator!(RawU1, 3);
-impl_bits_iterator!(RawU2, 2);
-impl_bits_iterator!(RawU4, 1);
-
-impl<'a, BO> IntoIterator for RawDataSlice<'a, RawU8, BO> {
-    type Item = RawU8;
-    type IntoIter = ByteIterator<'a>;
+impl<'a, R: RawData, O: DataOrder> IntoIterator for RawDataSlice<'a, R, O> {
+    type Item = R;
+    type IntoIter = RawDataIterator<'a, R, O>;
 
     fn into_iter(self) -> Self::IntoIter {
-        ByteIterator::new(self.data)
+        RawDataIterator::new(self.data)
     }
 }
 
-/// Iterator for raw data slices with less than 8 BPP.
-///
-/// See the [module-level documentation] for more information.
-///
-/// [module-level documentation]: self
+/// TODO
 #[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(::defmt::Format))]
-pub struct BitsIterator<'a, R> {
+pub struct RawDataIterator<'a, R, O> {
     data: &'a [u8],
     index: usize,
     raw_type: PhantomData<R>,
+    data_order: PhantomData<O>,
 }
 
-impl<'a, R: RawData> BitsIterator<'a, R> {
-    const fn new(data: &'a [u8]) -> Self {
+impl<'a, R, O> RawDataIterator<'a, R, O> {
+    fn new(data: &'a [u8]) -> Self {
         Self {
             data,
             index: 0,
             raw_type: PhantomData,
+            data_order: PhantomData,
         }
     }
 }
 
-/// Iterator for raw data slices with 8 BPP.
-///
-/// See the [module-level documentation] for more information.
-///
-/// [module-level documentation]: self
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(::defmt::Format))]
-pub struct ByteIterator<'a> {
-    data: slice::Iter<'a, u8>,
-}
+impl<R: RawData, O: DataOrder> Iterator for RawDataIterator<'_, R, O> {
+    type Item = R;
 
-impl<'a> ByteIterator<'a> {
-    fn new(data: &'a [u8]) -> Self {
-        Self { data: data.iter() }
-    }
-}
-
-impl<'a> Iterator for ByteIterator<'a> {
-    type Item = RawU8;
-
-    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
-        self.data.next().copied().map(RawU8::new)
+        R::load::<O>(self.data, self.index).map(|v| {
+            self.index += 1;
+            v
+        })
     }
 
     #[inline]
     fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.data.nth(n).copied().map(RawU8::new)
+        self.index = self.index.saturating_add(n);
+        self.next()
     }
 
     fn size_hint(&self) -> (usize, Option<usize>) {
-        self.data.size_hint()
+        let pixels_total = if R::BITS_PER_PIXEL >= 8 {
+            self.data.len() * (8 / R::BITS_PER_PIXEL)
+        } else {
+            self.data.len() * (R::BITS_PER_PIXEL / 8)
+        };
+
+        let size = pixels_total.saturating_sub(self.index);
+
+        (size, Some(size))
     }
 }
-
-/// Iterator for raw data slices more than 8 BPP.
-///
-/// See the [module-level documentation] for more information.
-///
-/// [module-level documentation]: self
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(::defmt::Format))]
-pub struct BytesIterator<'a, R, BO> {
-    // MSRV: replace by ArrayChunks when the feature is stabilized
-    data: slice::ChunksExact<'a, u8>,
-    raw_type: PhantomData<R>,
-    byte_order: PhantomData<BO>,
-}
-
-impl<'a, R: RawData, BO> BytesIterator<'a, R, BO> {
-    fn new(data: &'a [u8]) -> Self {
-        Self {
-            data: data.chunks_exact(R::BITS_PER_PIXEL / 8),
-            raw_type: PhantomData,
-            byte_order: PhantomData,
-        }
-    }
-}
-
-macro_rules! impl_bytes_iterator {
-    ($type:ident, $byte_order:ident, $read_function:path) => {
-        impl<'a> Iterator for BytesIterator<'a, $type, $byte_order> {
-            type Item = $type;
-
-            #[inline]
-            fn next(&mut self) -> Option<Self::Item> {
-                self.data.next().map(|raw| $read_function(raw).into())
-            }
-
-            #[inline]
-            fn nth(&mut self, n: usize) -> Option<Self::Item> {
-                self.data.nth(n).map(|raw| $read_function(raw).into())
-            }
-
-            fn size_hint(&self) -> (usize, Option<usize>) {
-                self.data.size_hint()
-            }
-        }
-
-        impl<'a> IntoIterator for RawDataSlice<'a, $type, $byte_order> {
-            type Item = $type;
-            type IntoIter = BytesIterator<'a, $type, $byte_order>;
-
-            fn into_iter(self) -> Self::IntoIter {
-                BytesIterator::new(self.data)
-            }
-        }
-    };
-
-    ($type:ident, $read_function:ident) => {
-        impl_bytes_iterator!($type, LittleEndian, LE::$read_function);
-        impl_bytes_iterator!($type, BigEndian, BE::$read_function);
-    };
-}
-
-impl_bytes_iterator!(RawU16, read_u16);
-impl_bytes_iterator!(RawU24, read_u24);
-impl_bytes_iterator!(RawU32, read_u32);
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::pixelcolor::raw::*;
 
     const BITS_DATA: &[u8] = &[0x12, 0x48, 0x5A, 0x0F];
     const BYTES_DATA_1: &[u8] = &[0x10, 0x20, 0x30, 0x40, 0x50, 0x60];
@@ -266,7 +140,7 @@ mod tests {
         .copied()
         .map(RawU1::new);
 
-        let iter = RawDataSlice::<RawU1, LittleEndian>::new(BITS_DATA).into_iter();
+        let iter = RawDataSlice::<RawU1, LittleEndianMsb0>::new(BITS_DATA).into_iter();
         assert!(iter.eq(expected));
     }
 
@@ -277,7 +151,7 @@ mod tests {
             .copied()
             .map(RawU2::new);
 
-        let iter = RawDataSlice::<RawU2, LittleEndian>::new(BITS_DATA).into_iter();
+        let iter = RawDataSlice::<RawU2, LittleEndianMsb0>::new(BITS_DATA).into_iter();
         assert!(iter.eq(expected));
     }
 
@@ -288,7 +162,7 @@ mod tests {
             .copied()
             .map(RawU4::new);
 
-        let iter = RawDataSlice::<RawU4, LittleEndian>::new(BITS_DATA).into_iter();
+        let iter = RawDataSlice::<RawU4, LittleEndianMsb0>::new(BITS_DATA).into_iter();
         assert!(iter.eq(expected));
     }
 
@@ -296,7 +170,7 @@ mod tests {
     fn raw_u8() {
         let expected = BYTES_DATA_1.iter().map(|&v| RawU8::new(v));
 
-        let iter = RawDataSlice::<RawU8, LittleEndian>::new(BYTES_DATA_1).into_iter();
+        let iter = RawDataSlice::<RawU8, LittleEndianMsb0>::new(BYTES_DATA_1).into_iter();
         assert!(iter.eq(expected));
     }
 
@@ -304,7 +178,7 @@ mod tests {
     fn raw_u16_le() {
         let expected = [0x2010, 0x4030, 0x6050].iter().copied().map(RawU16::new);
 
-        let iter = RawDataSlice::<RawU16, LittleEndian>::new(BYTES_DATA_1).into_iter();
+        let iter = RawDataSlice::<RawU16, LittleEndianMsb0>::new(BYTES_DATA_1).into_iter();
         assert!(iter.eq(expected));
     }
 
@@ -312,13 +186,13 @@ mod tests {
     fn raw_u16_be() {
         let expected = [0x1020, 0x3040, 0x5060].iter().copied().map(RawU16::new);
 
-        let iter = RawDataSlice::<RawU16, BigEndian>::new(BYTES_DATA_1).into_iter();
+        let iter = RawDataSlice::<RawU16, BigEndianLsb0>::new(BYTES_DATA_1).into_iter();
         assert!(iter.eq(expected));
     }
 
     #[test]
     fn raw_u16_excess_bytes_are_ignored() {
-        let iter = RawDataSlice::<RawU16, LittleEndian>::new(&[0; 3]).into_iter();
+        let iter = RawDataSlice::<RawU16, LittleEndianMsb0>::new(&[0; 3]).into_iter();
         assert_eq!(iter.count(), 1);
     }
 
@@ -326,7 +200,7 @@ mod tests {
     fn raw_u24_le() {
         let expected = [0x302010, 0x605040].iter().copied().map(RawU24::new);
 
-        let iter = RawDataSlice::<RawU24, LittleEndian>::new(BYTES_DATA_1).into_iter();
+        let iter = RawDataSlice::<RawU24, LittleEndianMsb0>::new(BYTES_DATA_1).into_iter();
         assert!(iter.into_iter().eq(expected));
     }
 
@@ -334,13 +208,13 @@ mod tests {
     fn raw_u24_be() {
         let expected = [0x102030, 0x405060].iter().copied().map(RawU24::new);
 
-        let iter = RawDataSlice::<RawU24, BigEndian>::new(BYTES_DATA_1).into_iter();
+        let iter = RawDataSlice::<RawU24, BigEndianLsb0>::new(BYTES_DATA_1).into_iter();
         assert!(iter.into_iter().eq(expected));
     }
 
     #[test]
     fn raw_u24_excess_bytes_are_ignored() {
-        let iter = RawDataSlice::<RawU24, LittleEndian>::new(&[0; 7]).into_iter();
+        let iter = RawDataSlice::<RawU24, LittleEndianMsb0>::new(&[0; 7]).into_iter();
         assert_eq!(iter.count(), 2);
     }
 
@@ -348,7 +222,7 @@ mod tests {
     fn raw_u32_le() {
         let expected = [0x40302010, 0x80706050].iter().copied().map(RawU32::new);
 
-        let iter = RawDataSlice::<RawU32, LittleEndian>::new(BYTES_DATA_2).into_iter();
+        let iter = RawDataSlice::<RawU32, LittleEndianMsb0>::new(BYTES_DATA_2).into_iter();
         assert!(iter.into_iter().eq(expected));
     }
 
@@ -356,13 +230,13 @@ mod tests {
     fn raw_u32_be() {
         let expected = [0x10203040, 0x50607080].iter().copied().map(RawU32::new);
 
-        let iter = RawDataSlice::<RawU32, BigEndian>::new(BYTES_DATA_2).into_iter();
+        let iter = RawDataSlice::<RawU32, BigEndianLsb0>::new(BYTES_DATA_2).into_iter();
         assert!(iter.into_iter().eq(expected));
     }
 
     #[test]
     fn raw_u32_excess_bytes_are_ignored() {
-        let iter = RawDataSlice::<RawU32, LittleEndian>::new(&[0; 13]).into_iter();
+        let iter = RawDataSlice::<RawU32, LittleEndianMsb0>::new(&[0; 13]).into_iter();
         assert_eq!(iter.count(), 3);
     }
 }